clean folder structure for frontend
