--- frontend/src/components/SignupPage.js
@@ -1,127 +1,0 @@
-import {
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  TextField,
-  Typography,
-} from "@mui/material";
-import { useState } from "react";
-import axios from "axios";
-import { URL_USER_SVC } from "../configs";
-import {
-  STATUS_CODE_BAD_REQUEST,
-  STATUS_CODE_CONFLICT,
-  STATUS_CODE_CREATED,
-} from "../constants";
-import { Link } from "react-router-dom";
-
-const LOGIN_URL = "/login";
-
-function SignupPage() {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [dialogTitle, setDialogTitle] = useState("");
-  const [dialogMsg, setDialogMsg] = useState("");
-  const [isSignupSuccess, setIsSignupSuccess] = useState(false);
-
-  const handleSignup = async () => {
-    setIsSignupSuccess(false);
-    const res = await axios
-      .post(URL_USER_SVC, { username, password })
-      .catch((err) => {
-        console.log(err.response);
-        if (err.response.status === STATUS_CODE_CONFLICT) {
-          setErrorDialog("This username already exists");
-        } else if (err.response.status === STATUS_CODE_BAD_REQUEST) {
-          setErrorDialog("Username and/or Password are missing!");
-        } else {
-          setErrorDialog("Please try again later");
-        }
-      });
-    if (res && res.status === STATUS_CODE_CREATED) {
-      setSuccessDialog("Account successfully created");
-      setIsSignupSuccess(true);
-    }
-  };
-
-  const closeDialog = () => setIsDialogOpen(false);
-
-  const setSuccessDialog = (msg) => {
-    setIsDialogOpen(true);
-    setDialogTitle("Success");
-    setDialogMsg(msg);
-  };
-
-  const setErrorDialog = (msg) => {
-    setIsDialogOpen(true);
-    setDialogTitle("Error");
-    setDialogMsg(msg);
-  };
-
-  return (
-    <Box
-      display={"flex"}
-      flexDirection={"column"}
-      width={"100%"}
-      maxWidth={"400px"}
-    >
-      <Typography variant={"h3"} marginBottom={"2rem"}>
-        Sign Up
-      </Typography>
-      <TextField
-        label="Username"
-        variant="standard"
-        required
-        value={username}
-        onChange={(e) => setUsername(e.target.value)}
-        sx={{ marginBottom: "1rem" }}
-        autoFocus
-      />
-      <TextField
-        label="Password"
-        variant="standard"
-        required
-        type="password"
-        value={password}
-        onChange={(e) => setPassword(e.target.value)}
-        sx={{ marginBottom: "2rem" }}
-      />
-      <Box
-        display={"flex"}
-        alignItems={"center"}
-        gap={"20px"}
-        flexWrap={"wrap"}
-      >
-        <Button variant={"outlined"} onClick={handleSignup}>
-          Sign up
-        </Button>
-        <Typography>
-          Already have an account? Please <Link to={LOGIN_URL}>login</Link>
-        </Typography>
-      </Box>
-
-      <Dialog open={isDialogOpen} onClose={closeDialog}>
-        <DialogTitle>{dialogTitle}</DialogTitle>
-        <DialogContent>
-          <DialogContentText>{dialogMsg}</DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          {isSignupSuccess ? (
-            <Button component={Link} to="/login">
-              Log in
-            </Button>
-          ) : (
-            <Button onClick={closeDialog}>Done</Button>
-          )}
-        </DialogActions>
-      </Dialog>
-    </Box>
-  );
-}
-
-export default SignupPage;
